+++ Test-MAME-YM2608/settings.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <vector>
+#include <QString>
 
 class Settings
 {
@@ -12,11 +13,9 @@ public:
 	void setRate(int rate);
 	int getDuration() const;
 	void setDuration(int duration);
-	void setinputOrder(std::vector<int> order);
-	std::vector<int> getInputOrder() const;
 
 private:
+	static const QString SETTINGS_PATH_;
 	int rate_;
 	int duration_;
-	std::vector<int> inputOrder_;
 };